feat: implement authentication error handling and loading states in auth pages
