+++ app/auth/auth-code-error/auth-code-error-content.tsx
@@ -1,11 +1,26 @@
-import { Suspense } from 'react'
+"use client"
+
+import { useSearchParams } from 'next/navigation'
 import { Card } from '@/components/ui/card'
 import { Button } from '@/components/ui/button'
 import { AlertTriangle } from 'lucide-react'
 import Link from 'next/link'
-import { AuthCodeErrorContent } from './auth-code-error-content'
 
-function LoadingFallback() {
+export function AuthCodeErrorContent() {
+  const searchParams = useSearchParams()
+  const error = searchParams.get('error')
+
+  const getErrorMessage = (errorCode: string | null) => {
+    switch (errorCode) {
+      case 'no-code':
+        return 'No se recibió código de autorización'
+      case 'unexpected':
+        return 'Error inesperado durante la autenticación'
+      default:
+        return errorCode || 'Error desconocido en la autenticación'
+    }
+  }
+
   return (
     <div className="min-h-screen flex items-center justify-center p-4">
       <Card className="w-full max-w-md p-6">
@@ -17,7 +32,7 @@ function LoadingFallback() {
           <div className="text-center space-y-2">
             <h1 className="text-xl font-semibold">Error de Autenticación</h1>
             <p className="text-sm text-muted-foreground">
-              Cargando información del error...
+              {getErrorMessage(error)}
             </p>
           </div>
 
@@ -37,13 +52,5 @@ function LoadingFallback() {
         </div>
       </Card>
     </div>
-  )
-}
-
-export default function AuthCodeErrorPage() {
-  return (
-    <Suspense fallback={<LoadingFallback />}>
-      <AuthCodeErrorContent />
-    </Suspense>
   )
 }
